src/requirements.txt
cv
numpy
scikit-image
matplotlib
torch
torchvision
opencv-python
Pillow